+++ test/HttpTaskManagerPrioritizedTest.java
@@ -16,10 +16,12 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import static util.GsonConfigurator.getGson;
+
 public class HttpTaskManagerPrioritizedTest {
     TaskManager manager = Managers.getDefault();
     HttpTaskServer taskServer = new HttpTaskServer(manager);
-    Gson gson = HttpTaskServer.getGson();
+    Gson gson = getGson();
 
     public HttpTaskManagerPrioritizedTest() {
     }
