+++ test/HttpTaskManagerSubtasksTest.java
@@ -18,11 +18,12 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static util.GsonConfigurator.getGson;
 
 public class HttpTaskManagerSubtasksTest {
     TaskManager manager = Managers.getDefault();
     HttpTaskServer taskServer = new HttpTaskServer(manager);
-    Gson gson = HttpTaskServer.getGson();
+    Gson gson = getGson();
 
     public HttpTaskManagerSubtasksTest() {
     }
